removes jsx from interpolate
